requirements.txt
streamlit
pandas
plotly
scikit-learn
google-genai
groq
python-dotenv
numpy
